take comment directives handling to adapter-utils
